+++ kits/python/agent.py
@@ -70,12 +70,12 @@ class Agent():
                     closest_ice_tile = ice_tile_locations[np.argmin(ice_tile_distances)]
                     if np.all(closest_ice_tile == unit.pos):
                         if unit.power >= unit.dig_cost(game_state) + unit.action_queue_cost(game_state):
-                            actions[unit_id] = [unit.dig(repeat=0)]
+                            actions[unit_id] = [unit.dig(repeat=0, n=1)]
                     else:
                         direction = direction_to(unit.pos, closest_ice_tile)
                         move_cost = unit.move_cost(game_state, direction)
                         if move_cost is not None and unit.power >= move_cost + unit.action_queue_cost(game_state):
-                            actions[unit_id] = [unit.move(direction, repeat=0)]
+                            actions[unit_id] = [unit.move(direction, repeat=0, n=1)]
                 # else if we have enough ice, we go back to the factory and dump it.
                 elif unit.cargo.ice >= 40:
                     direction = direction_to(unit.pos, closest_factory_tile)
@@ -85,5 +85,5 @@ class Agent():
                     else:
                         move_cost = unit.move_cost(game_state, direction)
                         if move_cost is not None and unit.power >= move_cost + unit.action_queue_cost(game_state):
-                            actions[unit_id] = [unit.move(direction, repeat=0)]
+                            actions[unit_id] = [unit.move(direction, repeat=0, n=1)]
         return actions