+++ luxai_s2/luxai_s2/actions.py
@@ -16,18 +16,16 @@ from luxai_s2.map.position import Position
 import luxai_s2.unit as luxai_unit
 
 
-# (0 = move, 1 = transfer X amount of R, 2 = pickup X amount of R, 3 = dig, 4 = self destruct, 5 = recharge X, 6 = repeat)
+# (0 = move, 1 = transfer X amount of R, 2 = pickup X amount of R, 3 = dig, 4 = self destruct, 5 = recharge X)
 class Action:
     def __init__(self, act_type: str) -> None:
         self.act_type = act_type
-        self.repeat = 0
+        self.n = 1 # number of times to execute the action
         self.power_cost = 0
+        self.repeat = False # whether to put action to the back of the action queue
 
     def state_dict(self):
         raise NotImplementedError("")
-    @property
-    def repeating(self):
-        return self.repeat == -1 or self.repeat > 0
 
 
 class FactoryBuildAction(Action):
@@ -52,74 +50,80 @@ class FactoryWaterAction(Action):
 
 
 class MoveAction(Action):
-    def __init__(self, move_dir: int, dist: int = 1, repeat=0) -> None:
+    def __init__(self, move_dir: int, dist: int = 1, repeat: bool = False, n: int = 1) -> None:
         super().__init__("move")
         # a[1] = direction (0 = center, 1 = up, 2 = right, 3 = down, 4 = left)
         self.move_dir = move_dir
         self.dist = dist
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([0, self.move_dir, 0, 0, self.repeat])
+        return np.array([0, self.move_dir, 0, 0, self.repeat, self.n])
 
 
 class TransferAction(Action):
-    def __init__(self, transfer_dir: int, resource: int, transfer_amount: int, repeat=0) -> None:
+    def __init__(self, transfer_dir: int, resource: int, transfer_amount: int, repeat: bool = False, n: int = 1) -> None:
         super().__init__("transfer")
         # a[2] = R = resource type (0 = ice, 1 = ore, 2 = water, 3 = metal, 4 power)
         self.transfer_dir = transfer_dir
         self.resource = resource
         self.transfer_amount = transfer_amount
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([1, self.transfer_dir, self.resource, self.transfer_amount, self.repeat])
+        return np.array([1, self.transfer_dir, self.resource, self.transfer_amount, self.repeat, self.n])
 
 
 class PickupAction(Action):
-    def __init__(self, resource: int, pickup_amount: int, repeat=0) -> None:
+    def __init__(self, resource: int, pickup_amount: int, repeat: bool = False, n: int = 1) -> None:
         super().__init__("pickup")
         # a[2] = R = resource type (0 = ice, 1 = ore, 2 = water, 3 = metal, 4 power)
         self.resource = resource
         self.pickup_amount = pickup_amount
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([2, 0, self.resource, self.pickup_amount, self.repeat])
+        return np.array([2, 0, self.resource, self.pickup_amount, self.repeat, self.n])
 
 
 class DigAction(Action):
-    def __init__(self, repeat=0) -> None:
+    def __init__(self, repeat: bool = False, n: int = 1) -> None:
         super().__init__("dig")
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([3, 0, 0, 0, self.repeat])
+        return np.array([3, 0, 0, 0, self.repeat, self.n])
 
 
 class SelfDestructAction(Action):
-    def __init__(self, repeat=0) -> None:
+    def __init__(self, repeat: bool = False, n: int = 1) -> None:
         super().__init__("self_destruct")
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([4, 0, 0, 0, self.repeat])
+        return np.array([4, 0, 0, 0, self.repeat, self.n])
 
 
 class RechargeAction(Action):
-    def __init__(self, power: int, repeat=0) -> None:
+    def __init__(self, power: int, repeat: bool = False, n: int = 1) -> None:
         super().__init__("recharge")
         self.power = power
         self.repeat = repeat
+        self.n = n
         self.power_cost = 0
 
     def state_dict(self):
-        return np.array([5, 0, 0, self.power, self.repeat])
+        return np.array([5, 0, 0, self.power, self.repeat, self.n])
 
 
 def format_factory_action(a: int):
@@ -138,17 +142,17 @@ def format_action_vec(a: np.ndarray):
     # (0 = move, 1 = transfer X amount of R, 2 = pickup X amount of R, 3 = dig, 4 = self destruct, 5 = recharge X)
     a_type = a[0]
     if a_type == 0:
-        return MoveAction(a[1], dist=1, repeat=a[4])
+        return MoveAction(a[1], dist=1, repeat=a[4], n=a[5])
     elif a_type == 1:
-        return TransferAction(a[1], a[2], a[3], repeat=a[4])
+        return TransferAction(a[1], a[2], a[3], repeat=a[4], n=a[5])
     elif a_type == 2:
-        return PickupAction(a[2], a[3], repeat=a[4])
+        return PickupAction(a[2], a[3], repeat=a[4], n=a[5])
     elif a_type == 3:
-        return DigAction(repeat=a[4])
+        return DigAction(repeat=a[4], n=a[5])
     elif a_type == 4:
-        return SelfDestructAction(repeat=a[4])
+        return SelfDestructAction(repeat=a[4], n=a[5])
     elif a_type == 5:
-        return RechargeAction(a[3], repeat=a[4])
+        return RechargeAction(a[3], repeat=a[4], n=a[5])
     else:
         raise ValueError(f"Action {a} is invalid type, {a[0]} is not valid")